Cell.java : Factory Method Pattern
Factory Method Pattern applied(Cell - Abstract, CellOriginal - Concrete).

Originally, Cell.java was concrete cell, but when it comes to implement new Cells that have different move or reproduction mechanism, Original code violate OCP.

So, I made Cell.java abstract class. And made CellOriginal.java extends Cell.java and  work exactly same way as Original Cell.java.

As a result, It doesn't make any violation of OCP when it comes to create new cells that have different features.
